chore(styleguide): move styleguidist to the .styleguide directory
