+++ packages/retail-ui/.gitignore
@@ -9,7 +9,7 @@ coverage/
 gemini-report
 /docs/dist/
 /build
-/styleguide
+.styleguide/build
 /storybook-static
 *.less.d.ts
 *.less.flow
