+++ packages/retail-ui/.styleguide/components/Pathline/PathlineRenderer.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Styled from 'react-styleguidist/lib/rsg-components/Styled';
-import Spinner from '../../components/Spinner';
-import Link from '../../components/Link';
-import Gapped from '../../components/Gapped';
-import fetch from '../../lib/net/fetch-cors';
+import Spinner from '../../../components/Spinner';
+import Link from '../../../components/Link';
+import Gapped from '../../../components/Gapped';
+import fetch from '../../../lib/net/fetch-cors';
 import OriginalPathlineRenderer from 'react-styleguidist/lib/rsg-components/Pathline/PathlineRenderer';
 
 interface GithubIssue {
