+++ packages/retail-ui/.styleguide/config.js
@@ -1,13 +1,13 @@
 const fs = require('fs');
 const path = require('path');
-const parseTsComponent = require('react-docgen-typescript').withCustomConfig('./tsconfig.json', {
+const parseTsComponent = require('react-docgen-typescript').withCustomConfig(path.join(__dirname, '../tsconfig.json'), {
   propFilter: prop => !(prop.parent && /node_modules/.test(prop.parent.fileName)),
 }).parse;
 const parseJsComponent = require('react-docgen').parse;
-const libraryVersion = require('./package.json').version;
+const libraryVersion = require('../package.json').version;
 
 function getComponentList() {
-  const dirPath = path.resolve(__dirname, 'components');
+  const dirPath = path.resolve(__dirname, '../components');
   return fs
     .readdirSync(dirPath)
     .map(x => [path.join(dirPath, x, `${x}.tsx`), path.join(dirPath, x, `${x}.js`)])
@@ -171,7 +171,7 @@ const webpackConfig = {
 
 module.exports = {
   components: getComponentList,
-  styleguideDir: path.resolve(__dirname, `styleguide/${libraryVersion}`),
+  styleguideDir: path.resolve(__dirname, `build/${libraryVersion}`),
   skipComponentsWithoutExample: true,
   pagePerSection: true,
   styles,
@@ -192,5 +192,8 @@ module.exports = {
   version: libraryVersion,
   ribbon: {
     url: 'https://github.com/skbkontur/retail-ui',
+  },
+  getComponentPathLine(path) {
+    return path.substring(path.indexOf('components')) || path;
   },
 };
