+++ packages/retail-ui/package.json
@@ -15,12 +15,12 @@
     "clean": "git clean -fdxqe node_modules",
     "build": "yarn tsc -p prod.tsconfig.json && node scripts/build",
     "postpublish": "yarn deploy && yarn publish:react-ui",
-    "predeploy": "rimraf styleguide && yarn run styleguide:build",
-    "deploy": "gh-pages -a -d styleguide -r git@github.com:skbkontur/react-ui.git",
+    "predeploy": "rimraf .styleguide/build && yarn run styleguide:build",
+    "deploy": "gh-pages -a -d .styleguide/build -r git@github.com:skbkontur/react-ui.git",
     "publish:react-ui": "yarn --cwd ./build publish --non-interactive",
     "storybook": "start-storybook -p 6060",
-    "styleguide": "styleguidist server",
-    "styleguide:build": "styleguidist build && styleguidist build --config styleguide.prod.config.js",
+    "styleguide": "styleguidist server --config .styleguide/config.js",
+    "styleguide:build": "styleguidist build --config .styleguide/config.js && styleguidist build --config .styleguide/prod.config.js",
     "lint:flow": "flow check",
     "lint": "npm-run-all --continue-on-error --parallel lint:*",
     "lint:tsc": "tsc --noEmit",
