+++ packages/retail-ui/prod.tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["**/__mocks__/**/*", "**/__tests__/**/*", "**/__stories__/**/*", "./styleguide", "./scripts"]
+  "exclude": ["**/__mocks__/**/*", "**/__tests__/**/*", "**/__stories__/**/*", ".styleguide"]
 }
